requirements.txt
alembic==1.3.1
aniso8601==8.0.0
Click==7.0
Flask==1.1.1
Flask-API==2.0
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.19.0
mysqlclient==1.3.12
python-dateutil==2.8.1
python-decouple==3.3
python-editor==1.0.4
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.11
Werkzeug==0.16.0
psycopg2==2.8.4